Tested read_mods and group_by_seq
