+++ Test/Test.csproj
@@ -113,7 +113,6 @@
     <Compile Include="TestProteoformCommunityRelate.cs" />
     <Compile Include="TestCytoscapeScript.cs" />
     <Compile Include="TestProteoformFamilies.cs" />
-    <Compile Include="TestTheoreticalCreation.cs" />
     <Compile Include="TestTheoreticalDatabaseCreate.cs" />
   </ItemGroup>
   <ItemGroup>
@@ -141,6 +140,9 @@
     <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
   </ItemGroup>
   <ItemGroup>
+    <Content Include="amino_acids_duplicates.txt">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </Content>
     <Content Include="ptmlist.txt">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>
